refactor(git-hosts): extract the managed-workspace subsystem from managed_hosts
managed_hosts.py had grown a ~550-line git-repository subsystem (repo-URL
parsing, host resolution, per-owner credential-slot selection, relaunch
re-authorization, config hashing, clone-env building, binding labels) inside
sandbox-launcher lifecycle code. Move it verbatim into a focused
git_hosts/managed_workspace module (cycle-safe: no runtime server/store
imports, the store type is TYPE_CHECKING-only) so managed_hosts orchestrates
launchers and importers just change their import line. Also drop the
write-only MANAGED_GIT_CANONICAL_URL label, which nothing read. Fail-closed
selection, relaunch gates and their exact messages, and host_config_hash are
byte-identical.
